+++ CHANGELOG.md
@@ -21,3 +21,13 @@ Change log for **Simple-Router-PHP**
 
 ### Changed
 * The `endpoint` function was modified to make it private and use the new functions.
+
+## [1.0.2] - 2024-04-26
+
+### Added
+* Added support for grouping routes with common prefix path or middleware using the `group` method.
+
+## [1.0.3] - 2024-05-09
+
+### Changed
+* The `endpoint` function was modified to make it private again and the type of the options is now typed as an array.